Rag/requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
streamlit==1.31.1
faiss-cpu==1.7.4
pypdf==3.17.4
